toBalancesKey moved
